validacion url a la hora de checkout

+++ client/src/Components/CheckoutSuccess/CheckoutSuccess.jsx
@@ -5,13 +5,17 @@ import { Dispatch } from "react";
 import { useDispatch } from "react-redux";
 import { useSelector } from "react-redux";
 import { CreateOrder } from "../../Redux/actions";
+
+import { useLocation } from "react-router-dom";
+
 const CheckoutSuccess = () => {
 
 
   // -usuario toda la informacion
   // carrito
   // precio total del carrito
-
+  const location = useLocation();
+  console.log(location)
   const dispatch = useDispatch();
 
   const [usuario, saveUser] = useLocalStorage("USUARIO", " ");
@@ -26,10 +30,12 @@ const CheckoutSuccess = () => {
   let precio = totalPrice(item)
 
   
-  // "{"carrito":[{"title":"Zapatillas Para Hombre Jaguar 4325 Color Verde - Adulto 43 Ar","id":"ecad0881-9dee-4d33-bd3d-0c07ba6026df","price":9819,"sold_quantity":5,"condition":"new","pictures":["http://http2.mlstatic.com/D_800260-MLA49260581435_032022-O.jpg","http://http2.mlstatic.com/D_648638-MLA48346156852_112021-O.jpg","http://http2.mlstatic.com/D_707742-MLA48346254494_112021-O.jpg"],"available_quantity":1,"visible":0,"attributes":[{"id":8,"name":"Edad","value":"Adultos","productId":"ecad0881-9dee-4d33-bd3d-0c07ba6026df"},{"id":9,"name":"Marca","value":"Jaguar","productId":"ecad0881-9dee-4d33-bd3d-0c07ba6026df"},{"id":11,"name":"Materiales del exterior","value":"Tela","productId":"ecad0881-9dee-4d33-bd3d-0c07ba6026df"},{"id":12,"name":"Género","value":"Hombre","productId":"ecad0881-9dee-4d33-bd3d-0c07ba6026df"},{"id":13,"name":"Talle","value":"43 AR","productId":"ecad0881-9dee-4d33-bd3d-0c07ba6026df"},{"id":14,"name":"Estilo","value":"Deportivo","productId":"ecad0881-9dee-4d33-bd3d-0c07ba6026df"},{"id":10,"name":"Color","value":"Verde","productId":"ecad0881-9dee-4d33-bd3d-0c07ba6026df"}],"variations":[],"reviews":[],"cantidad":1}],"precioTotal":9819,"usuarioId":"5d01647a-8f3d-4a47-a114-243a1f79f532"}"
+
   useEffect(() => {
-    dispatch(CreateOrder({carrito: item, precioTotal: precio, usuarioId: usuario.id}))
-    localStorage.clear()
+    if(location.search === "?paid=true") {
+      dispatch(CreateOrder({carrito: item, precioTotal: precio, usuarioId: usuario.id}))
+      localStorage.clear()
+    }
   }, [])
 
   return (
